+++ GestioneCorsi/src/it/betacom/milano/businesscomponent/model/Corso_Corsista.java
@@ -1,22 +1,23 @@
 package it.betacom.milano.businesscomponent.model;
 
 public class Corso_Corsista {
-	private int codice_corso;
-	private int codice_corsista;
+	private long codice_corso;
+	private long codice_corsista;
 	
 	
 	
 	public Corso_Corsista() {}
-	public int getCodice_corso() {
+	
+	public long getCodice_corso() {
 		return codice_corso;
 	}
-	public void setCodice_corso(int codice_corso) {
+	public void setCodice_corso(long codice_corso) {
 		this.codice_corso = codice_corso;
 	}
-	public int getCodice_corsista() {
+	public long getCodice_corsista() {
 		return codice_corsista;
 	}
-	public void setCodice_corsista(int codice_corsista) {
+	public void setCodice_corsista(long codice_corsista) {
 		this.codice_corsista = codice_corsista;
 	}
